apps/ingestion-worker/requirements.txt
# Runtime
redis==5.2.1
openai==1.75.0
pymilvus==2.5.8
pymilvus[model]==2.5.8
boto3==1.37.20
psycopg2-binary==2.9.10
bm25s==0.2.12

# Extractors
pypdf==5.4.0
python-docx==1.1.2
beautifulsoup4==4.13.3
lxml==5.4.0
markdown==3.7
